WIP: Got much closer by skipping the padding if it's the last field in the format.  I got the idea the spec, although it's still a bit unclear to me if it's right.
It does still crash, but much further on now.

The [spec](https://github.com/PX4/PX4-Devguide/blob/master/en/log/ulog_file_format.md), on padding:

```
Padding: field names that start with _padding should not be displayed and their data must be ignored by a reader. These fields can be inserted by a writer to ensure correct alignment.

If the padding field is the last field, then this field will not be logged, to avoid writing unnecessary data. This means the message_data_s.data will be shorter by the size of the padding. However the padding is still needed when the message is used in a nested definition.

```

Also of note.

⚠️The PlotJuggler code makes an unsupported assumption that the timestamp is always the first field and they read it in `parse_data_message()` before calling `parse_simple_data_message()`.
For now we ket this convention in this code, but we'll have to refactothat to fix it.

This might explain why their code crashes.