Replaced the crap code in `CsvExporter::to_csv()`.  Works for simple examples and runs very fast. 🔥
The general algorithm is to first find the column with the lowest timestamp `t`, and then to search across all other columns for those which also contain data at `t`.
 All other values are replaced with empty values in the output.
It uses two methods:

* `CsvExporter::min_timestamp()` which finds the minimum current timestamp across all the series.
* `Column::value_at_time()` which returns the current data value at time `t` for a column, or `None` if no data is present.

The result is a CSV row at time `t` with only the data values filled in.  Typically not all topics in a PX$ ULOG file
will have the same timestamps, this typeically results in a sparse CSV file.