Introduce `#[ulog_preamble]` macro
Added a new attribute macro `#[ulog_preamble]` that wraps enums and structs into a module and
automatically implements the `FromField` trait for scalar and array types. Updated examples
and tests to place all `ULogData` and `ULogMessages` derives inside a `pub mod` annotated
with the macro.  This may be a way to avoid polluting the crate with definitions, and especially
 to allow for multiple enums annotated with  #[derive(ULogMessages)] to be defined.

I'm not sure whether I will retain this approach.  I may look at how serde handles it instead:

https://www.reddit.com/r/rust/comments/yypj09/deleted_by_user/